move dbTypes to type dir
