+++ src/db/models/__base_model.ts
@@ -1,5 +1,5 @@
 import * as sqlite from 'better-sqlite3';
-import * as types from '../dbTypes';
+import * as types from '../../types/dbTypes';
 
 export class Model<M> {
   protected static db: sqlite.Database;
