+++ src/db/models/usermatch_model.ts
@@ -1,5 +1,5 @@
 import * as sqlite from 'better-sqlite3';
-import * as types from '../dbTypes';
+import * as types from '../../types/dbTypes';
 import { Model } from './__base_model';
 import { UserModel, MatchModel } from '.';
